# Conflicts:
#	2. BackEnd/milkTea-server/src/main/java/com/vti/controller/OrderDetailController.java
#	2. BackEnd/milkTea-server/src/main/java/com/vti/dto/OrderDetailDTO.java

+++ 2. BackEnd/milkTea-server/src/main/java/com/vti/repository/IOrderDetailRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 public interface IOrderDetailRepository extends JpaRepository<OrderDetails, Integer>, JpaSpecificationExecutor<OrderDetails> {
-    boolean existsById(Long id);
+    boolean existsById(int id);
 
-    void deleteById(Long id);
+    void deleteById(int id);
 
 }
